requirements.txt
fastapi
uvicorn
openai
google-api-python-client
google-auth
sendgrid
python-dotenv
requests
google-auth-httplib2
google-auth-oauthlib
